Add files via upload (#1)
* Add files via upload

* Update clientLinux.js

* Linux directory

Specificly working Client and Server JS files that work on Linux (mint KDE and Cinnamon).

* Server side JS

The server side JS file. Run this on the computer you want to use as a controller.
Eventually this code will be running on a raspberry pi Wi-Pi-Dock.

* Delete clientLinux.js

Outdated. And oh so very bad.

* Delete serverLinux.js

Outdated and very, very bad.